Edited several files to address the assignment.
I added a placeholder "calculate_price" method for the pricing module. I made significant changes to views.py and fuelquote.html to pull the suggested price from the pricing module and multiply that by the number of gallons to get the total price. I also changed the name of the "QuoteHistory" model to the "Quote" model for clarity. This change should be reflected in all the files. Finally, I edited the "Address" model so that the "address2" field is no longer required because not all addresses have a second address line. To make referencing an address easier, I created a "__str__" function for that model to return a string of the address, accounting for addresses that have a second address line and ones that do not. I subsequently replaced any instance of converting an "Address" object to a string with a call to that function. There is a coverage report, but it is currently not meaningful. After all these changes, I ran migrate and makemigrate on the project.